Split of tasks API into its own namespace.
