+++ spec/common/providers/https.spec.ts
@@ -1,20 +1,18 @@
 import { expect } from 'chai';
-import * as express from 'express';
 import * as firebase from 'firebase-admin';
 import * as sinon from 'sinon';
 
 import { apps as appsNamespace } from '../../../src/apps';
-import * as debug from '../../../src/common/debug';
-import * as https from '../../../src/common/providers/https';
 import {
-  CallableContext,
-  CallableRequest,
-  TaskContext,
-  TaskRequest,
+  checkAppCheckContext,
+  checkAuthContext,
+  runHandler,
+  RunHandlerResult,
+} from '../../helper';
+import {
   unsafeDecodeAppCheckToken,
   unsafeDecodeIdToken,
 } from '../../../src/common/providers/https';
-import * as mocks from '../../fixtures/credential/key.json';
 import {
   expectedResponseHeaders,
   generateAppCheckToken,
@@ -25,15 +23,9 @@ import {
   mockFetchPublicKeys,
   mockRequest,
 } from '../../fixtures/mockrequest';
-
-/**
- * RunHandlerResult contains the data from an express.Response.
- */
-interface RunHandlerResult {
-  status: number;
-  headers: { [name: string]: string };
-  body: any;
-}
+import * as debug from '../../../src/common/debug';
+import * as https from '../../../src/common/providers/https';
+import * as mocks from '../../fixtures/credential/key.json';
 
 /**
  * A CallTest is a specification for a test of a callable function that
@@ -56,64 +48,6 @@ interface CallTest {
 
   // The expected shape of the http response returned to the callable SDK.
   expectedHttpResponse: RunHandlerResult;
-}
-
-/**
- * Runs an express handler with a given request asynchronously and returns the
- * data populated into the response.
- */
-function runHandler(
-  handler: express.Handler,
-  request: https.Request
-): Promise<RunHandlerResult> {
-  return new Promise((resolve, reject) => {
-    // MockResponse mocks an express.Response.
-    // This class lives here so it can reference resolve and reject.
-    class MockResponse {
-      private statusCode = 0;
-      private headers: { [name: string]: string } = {};
-      private callback: Function;
-
-      public status(code: number) {
-        this.statusCode = code;
-        return this;
-      }
-
-      // Headers are only set by the cors handler.
-      public setHeader(name: string, value: string) {
-        this.headers[name] = value;
-      }
-
-      public getHeader(name: string): string {
-        return this.headers[name];
-      }
-
-      public send(body: any) {
-        resolve({
-          status: this.statusCode,
-          headers: this.headers,
-          body,
-        });
-        if (this.callback) {
-          this.callback();
-        }
-      }
-
-      public end() {
-        this.send(undefined);
-      }
-
-      public on(event: string, callback: Function) {
-        if (event !== 'finish') {
-          throw new Error('MockResponse only implements the finish event');
-        }
-        this.callback = callback;
-      }
-    }
-
-    const response = new MockResponse();
-    handler(request, response as any, () => undefined);
-  });
 }
 
 // Runs a CallTest test.
@@ -143,81 +77,6 @@ async function runCallableTest(test: CallTest): Promise<any> {
   expect(responseV2.body).to.deep.equal(test.expectedHttpResponse.body);
   expect(responseV2.headers).to.deep.equal(test.expectedHttpResponse.headers);
   expect(responseV2.status).to.equal(test.expectedHttpResponse.status);
-}
-
-/** Represents a test case for a Task Queue Function */
-interface TaskTest {
-  // An http request, mocking a subset of https.Request.
-  httpRequest: any;
-
-  // The expected format of the request passed to the handler.
-  expectedData: any;
-
-  taskFunction?: (
-    data: any,
-    context: https.TaskContext
-  ) => void | Promise<void>;
-
-  taskFunction2?: (request: https.TaskRequest<any>) => void | Promise<void>;
-
-  // The expected shape of the http response returned to the callable SDK.
-  expectedStatus: number;
-}
-
-// Runs a TaskTest test.
-async function runTaskTest(test: TaskTest): Promise<any> {
-  const taskQueueFunctionV1 = https.onDispatchHandler((data, context) => {
-    expect(data).to.deep.equal(test.expectedData);
-    if (test.taskFunction) {
-      test.taskFunction(data, context);
-    }
-  });
-
-  const responseV1 = await runHandler(taskQueueFunctionV1, test.httpRequest);
-  expect(responseV1.status).to.equal(test.expectedStatus);
-
-  const taskQueueFunctionV2 = https.onDispatchHandler((request) => {
-    expect(request.data).to.deep.equal(test.expectedData);
-    if (test.taskFunction2) {
-      test.taskFunction2(request);
-    }
-  });
-
-  const responseV2 = await runHandler(taskQueueFunctionV2, test.httpRequest);
-  expect(responseV2.status).to.equal(test.expectedStatus);
-}
-
-function checkAuthContext(
-  context: CallableContext | CallableRequest | TaskContext | TaskRequest,
-  projectId: string,
-  userId: string
-) {
-  expect(context.auth).to.not.be.undefined;
-  expect(context.auth).to.not.be.null;
-  expect(context.auth.uid).to.equal(userId);
-  expect(context.auth.token.uid).to.equal(userId);
-  expect(context.auth.token.sub).to.equal(userId);
-  expect(context.auth.token.aud).to.equal(projectId);
-
-  // TaskContext & TaskRequest don't have instanceIdToken
-  if ({}.hasOwnProperty.call(context, 'instanceIdToken')) {
-    expect((context as CallableContext).instanceIdToken).to.be.undefined;
-  }
-}
-
-function checkAppCheckContext(
-  context: CallableContext | CallableRequest,
-  projectId: string,
-  appId: string
-) {
-  expect(context.app).to.not.be.undefined;
-  expect(context.app).to.not.be.null;
-  expect(context.app.appId).to.equal(appId);
-  expect(context.app.token.app_id).to.be.equal(appId);
-  expect(context.app.token.sub).to.be.equal(appId);
-  expect(context.app.token.aud).to.be.deep.equal([`projects/${projectId}`]);
-  expect(context.auth).to.be.undefined;
-  expect(context.instanceIdToken).to.be.undefined;
 }
 
 describe('onCallHandler', () => {
@@ -662,187 +521,6 @@ describe('onCallHandler', () => {
           headers: expectedResponseHeaders,
           body: { result: null },
         },
-      });
-    });
-  });
-});
-
-describe('onEnqueueHandler', () => {
-  let app: firebase.app.App;
-
-  before(() => {
-    const credential = {
-      getAccessToken: () => {
-        return Promise.resolve({
-          expires_in: 1000,
-          access_token: 'fake',
-        });
-      },
-      getCertificate: () => {
-        return {
-          projectId: 'aProjectId',
-        };
-      },
-    };
-    app = firebase.initializeApp({
-      projectId: 'aProjectId',
-      credential,
-    });
-    Object.defineProperty(appsNamespace(), 'admin', { get: () => app });
-  });
-
-  after(() => {
-    app.delete();
-    delete appsNamespace.singleton;
-  });
-
-  it('should handle success', () => {
-    return runTaskTest({
-      httpRequest: mockRequest({ foo: 'bar' }),
-      expectedData: { foo: 'bar' },
-      expectedStatus: 204,
-    });
-  });
-
-  it('should reject bad method', () => {
-    const req = mockRequest(null);
-    req.method = 'GET';
-    return runTaskTest({
-      httpRequest: req,
-      expectedData: null,
-      expectedStatus: 400,
-    });
-  });
-
-  it('should ignore charset', () => {
-    return runTaskTest({
-      httpRequest: mockRequest(null, 'application/json; charset=utf-8'),
-      expectedData: null,
-      expectedStatus: 204,
-    });
-  });
-
-  it('should reject bad content type', () => {
-    return runTaskTest({
-      httpRequest: mockRequest(null, 'text/plain'),
-      expectedData: null,
-      expectedStatus: 400,
-    });
-  });
-
-  it('should reject extra body fields', () => {
-    const req = mockRequest(null);
-    req.body.extra = 'bad';
-    return runTaskTest({
-      httpRequest: req,
-      expectedData: null,
-      expectedStatus: 400,
-    });
-  });
-
-  it('should handle unhandled error', () => {
-    return runTaskTest({
-      httpRequest: mockRequest(null),
-      expectedData: null,
-      taskFunction: (data, context) => {
-        throw new Error(`ceci n'est pas une error`);
-      },
-      taskFunction2: (request) => {
-        throw new Error(`cece n'est pas une error`);
-      },
-      expectedStatus: 500,
-    });
-  });
-
-  it('should handle unknown error status', () => {
-    return runTaskTest({
-      httpRequest: mockRequest(null),
-      expectedData: null,
-      taskFunction: (data, context) => {
-        throw new https.HttpsError('THIS_IS_NOT_VALID' as any, 'nope');
-      },
-      taskFunction2: (request) => {
-        throw new https.HttpsError('THIS_IS_NOT_VALID' as any, 'nope');
-      },
-      expectedStatus: 500,
-    });
-  });
-
-  it('should handle well-formed error', () => {
-    return runTaskTest({
-      httpRequest: mockRequest(null),
-      expectedData: null,
-      taskFunction: (data, context) => {
-        throw new https.HttpsError('not-found', 'i am error');
-      },
-      taskFunction2: (request) => {
-        throw new https.HttpsError('not-found', 'i am error');
-      },
-      expectedStatus: 404,
-    });
-  });
-
-  it('should handle auth', async () => {
-    const mock = mockFetchPublicKeys();
-    const projectId = appsNamespace().admin.options.projectId;
-    const idToken = generateIdToken(projectId);
-    await runTaskTest({
-      httpRequest: mockRequest(null, 'application/json', {
-        authorization: 'Bearer ' + idToken,
-      }),
-      expectedData: null,
-      taskFunction: (data, context) => {
-        checkAuthContext(context, projectId, mocks.user_id);
-        return null;
-      },
-      taskFunction2: (request) => {
-        checkAuthContext(request, projectId, mocks.user_id);
-        return null;
-      },
-      expectedStatus: 204,
-    });
-    mock.done();
-  });
-
-  it('should reject bad auth', async () => {
-    const projectId = appsNamespace().admin.options.projectId;
-    const idToken = generateUnsignedIdToken(projectId);
-    await runTaskTest({
-      httpRequest: mockRequest(null, 'application/json', {
-        authorization: 'Bearer ' + idToken,
-      }),
-      expectedData: null,
-      expectedStatus: 401,
-    });
-  });
-
-  describe('skip token verification debug mode support', () => {
-    before(() => {
-      sinon
-        .stub(debug, 'isDebugFeatureEnabled')
-        .withArgs('skipTokenVerification')
-        .returns(true);
-    });
-
-    after(() => {
-      sinon.verifyAndRestore();
-    });
-
-    it('should skip auth token verification', async () => {
-      const projectId = appsNamespace().admin.options.projectId;
-      const idToken = generateUnsignedIdToken(projectId);
-      await runTaskTest({
-        httpRequest: mockRequest(null, 'application/json', {
-          authorization: 'Bearer ' + idToken,
-        }),
-        expectedData: null,
-        taskFunction: (data, context) => {
-          checkAuthContext(context, projectId, mocks.user_id);
-        },
-        taskFunction2: (request) => {
-          checkAuthContext(request, projectId, mocks.user_id);
-        },
-        expectedStatus: 204,
       });
     });
   });
