+++ src/function-builder.ts
@@ -42,6 +42,7 @@ import * as https from './providers/https';
 import * as pubsub from './providers/pubsub';
 import * as remoteConfig from './providers/remoteConfig';
 import * as storage from './providers/storage';
+import * as tasks from './providers/tasks';
 import * as testLab from './providers/testLab';
 
 /**
@@ -373,8 +374,8 @@ export class FunctionBuilder {
        * @param options Configurations for the task queue function.
        */
       /** @hidden */
-      taskQueue: (options?: https.TaskQueueOptions) => {
-        return new https.TaskQueueBuilder(options, this.options);
+      taskQueue: (options?: tasks.TaskQueueOptions) => {
+        return new tasks.TaskQueueBuilder(options, this.options);
       },
     };
   }
